[ch107] Move each form to each screen
