+++ src/components/SignInForm/index.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
-import Form from '../Form'
+import Form from './Form'
 
 const SignInForm = () => {
   return (
     <section className="sign-in-form">
       <header className="sign-in-form__header">Sign In</header>
       <div className="sign-in-form__body">
-        <Form title="Personal Detail" href="/" />
+        <Form />
       </div>
       <footer className="sign-in-form__footer">
         Don't have an account? Create here
